refactor: adapter 클래스들 및 바인딩어댑터를 adapter 하위 패키지로 이동
