+++ app/src/main/java/com/moroom/android/ResultActivity.java
@@ -23,6 +23,7 @@ import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
+import com.moroom.android.ui.adapter.home.BestReviewAdapter;
 
 import java.util.ArrayList;
