+++ app/src/main/java/com/moroom/android/WriteActivity.java
@@ -24,6 +24,7 @@ import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
+import com.moroom.android.ui.adapter.write.CheckedTextViewAdapter;
 
 import java.util.ArrayList;
 import java.util.function.Consumer;
