+++ app/src/main/java/com/moroom/android/ui/adapter/BindingAdapter.kt
@@ -1,10 +1,11 @@
-package com.moroom.android
+package com.moroom.android.ui.adapter
 
 import android.view.View
 import android.widget.ProgressBar
 import androidx.databinding.BindingAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.moroom.android.data.model.Review
+import com.moroom.android.ui.adapter.home.BestReviewAdapter
 
 object ReviewBindingAdapter {
     @JvmStatic
