+++ app/src/main/java/com/moroom/android/ui/navui/home/HomeFragment.kt
@@ -15,7 +15,6 @@ import androidx.recyclerview.widget.LinearLayoutManager
 
 import com.google.firebase.auth.FirebaseAuth
 
-import com.moroom.android.BestReviewViewModel
 import com.moroom.android.MoveToLogin
 import com.moroom.android.R
 import com.moroom.android.ResultActivity
@@ -26,14 +25,14 @@ import com.moroom.android.databinding.FragmentHomeBinding
 class HomeFragment : Fragment() {
     private var _binding: FragmentHomeBinding? = null
     private val binding get() = _binding!!
-    private lateinit var viewModel: BestReviewViewModel
+    private lateinit var viewModel: HomeViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        viewModel = ViewModelProvider(requireActivity()).get(BestReviewViewModel::class.java)
+        viewModel = ViewModelProvider(requireActivity()).get(HomeViewModel::class.java)
         _binding = DataBindingUtil.inflate(inflater, R.layout.fragment_home, container, false)
         binding.viewModel = viewModel
         binding.lifecycleOwner = viewLifecycleOwner
