--- app/src/main/java/com/moroom/android/ui/navui/home/HomeViewModel.java
@@ -1,17 +1,0 @@
-package com.moroom.android.ui.navui.home;
-
-import androidx.lifecycle.ViewModel;
-
-public class HomeViewModel extends ViewModel {
-
-//    private final MutableLiveData<String> mText;
-//
-//    public HomeViewModel() {
-//        mText = new MutableLiveData<>();
-//        mText.setValue("This is home fragment");
-//    }
-//
-//    public LiveData<String> getText() {
-//        return mText;
-//    }
-}
