+++ app/src/main/res/layout/fragment_home.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="viewModel"
-            type="com.moroom.android.BestReviewViewModel" />
+            type="com.moroom.android.ui.navui.home.HomeViewModel" />
     </data>
 
     <RelativeLayout
